Feat add scroll to filters panel (#29)
* feat add scroll to filters painel


* fix handleAddProtocol functionality


* feat add scroll to all filter sections


* style fix formatting


---------
